requirements.txt
alembic==1.2.1
aniso8601==8.0.0
astroid==2.3.0
attrs==19.1.0
awsebcli==3.15.3
bcrypt==3.1.7
beautifulsoup4==4.8.0
blessed==1.15.0
botocore==1.12.243
cached-property==1.5.1
cement==2.8.2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.3.9
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
future==0.16.0
idna==2.7
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==2.6.0
lazy-object-proxy==1.4.2
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.5.9
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.1
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
soupsieve==1.9.4
SQLAlchemy==1.3.9
termcolor==1.1.0
texttable==0.9.1
typed-ast==1.4.0
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.16.0
wrapt==1.11.2
